Implemented background screen with metaballs.
