+++ src/Game/Frenzied.XNA.Windows.csproj
@@ -78,7 +78,6 @@
   <ItemGroup>
     <Compile Include="Assets\AssetManager.cs" />
     <Compile Include="Audio\AudioManager.cs" />
-    <Compile Include="Components\Background.cs" />
     <Compile Include="Config\BackgroundConfig.cs" />
     <Compile Include="Config\GameConfig.cs" />
     <Compile Include="FrenziedGame.cs" />
@@ -95,10 +94,10 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Screens\GameScreen.cs" />
+    <Compile Include="Screens\Implementations\BackgroundScreen.cs" />
     <Compile Include="Screens\InputState.cs" />
     <Compile Include="Screens\ScreenManager.cs" />
     <Compile Include="ScreenOld\Implementations\GamePlayScreenOld.cs" />
-    <Compile Include="Screens\Implementations\BackgroundScreen.cs" />
     <Compile Include="Screens\Implementations\GameplayScreen.cs" />
     <Compile Include="Screens\Implementations\LoadingScreen.cs" />
     <Compile Include="Screens\Implementations\MainMenuScreen.cs" />
